flask/requirements.txt
Click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
flask-cors
psycopg2-binary==2.9.1
urllib3
